+++ docs/changelog.md
@@ -15,6 +15,7 @@ Includes several critical bugfixes and improvements when processing multiple inp
 
  * [general] Minimum supported Python version is now 3.10 (Python 3.9 is [EOL](https://devguide.python.org/versions/))
  * [general] An error is now displayed when launching the app fails due to an invalid config file [#241](https://github.com/Breakthrough/DVR-Scan/issues/241)
+ * [general] Python package now only depends on `opencv-contrib-python` or `opencv-contrib-python-headless` [#251](https://github.com/Breakthrough/DVR-Scan/issues/251)
  * [bugfix] Fix `quiet-mode` setting (`-q`/`--quiet` flag) still allowing extraneous output [#243](https://github.com/Breakthrough/DVR-Scan/issues/243)
  * [bugfix] Fix loading config from file in UI [#253](https://github.com/Breakthrough/DVR-Scan/issues/253)
  * [bugfix] Fix scan error when enabling "Use PTS" input option with multiple videos [#254](https://github.com/Breakthrough/DVR-Scan/issues/254)